Renamed session_serializer option to cookies_serializer
