--- actionpack/lib/action_dispatch/middleware/session/marshal_serializer.rb
@@ -1,14 +1,0 @@
-module ActionDispatch
-  module Session
-    class MarshalSerializer
-      def self.load(value)
-        Marshal.load(value)
-      end
-
-      def self.dump(value)
-        Marshal.dump(value)
-      end
-    end
-  end
-end
-
